client/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
click==8.1.7
exceptiongroup==1.1.3
fastapi==0.104.1
grpcio==1.59.3
grpcio-tools==1.59.3
h11==0.14.0
idna==3.4
protobuf==4.25.1
pydantic==2.5.1
pydantic_core==2.14.3
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.24.0.post1
